refactor: chart examples CSS and webpack structural changes
