+++ app/examples/chart/area-series-page.ts
