+++ app/examples/chart/bar-series-page.css
