+++ app/examples/chart/bar-series-page.ts
