+++ app/examples/chart/line-series-page.css
