+++ app/examples/chart/line-series-page.xml
@@ -1,11 +1,10 @@
 <examples:ExamplePage
-        cssFile="~/examples/chart/bar-series.css"
         navigatingTo="pageNavigatingTo"
         navigatingFrom="pageNavigatingFrom"
         xmlns:examples="examples/example-base-page"
         xmlns="http://www.nativescript.org/tns.xsd">
     <Page.actionBar>
-        <ActionBar title="Bar series" automationText="ActionBar">
+        <ActionBar title="Line series" automationText="ActionBar">
             <NavigationButton icon="res://ic_back" tap="goBack" automationText="GoBack" />
             <Android>
                 <ActionItem id="exampleMenuButton" icon="res://ic_menu_example" automationText="ExampleMenu" />
@@ -23,7 +22,7 @@
         <GridLayout class="repeaterItemsScrollerStyle" rows="*">
             <GridLayout class="repeaterBottomBorderStyle" row="1"/>
             <ScrollView id="scrollView" orientation="horizontal" ios.showsHorizontalScrollIndicator="false" class="galleryRepeaterStyle" loaded="scrollViewLoaded">
-                <Repeater items="{{ barTypes }}" >
+                <Repeater items="{{ lineTypes }}" >
                     <Repeater.itemsLayout>
                         <StackLayout orientation="horizontal"/>
                     </Repeater.itemsLayout>
