--- app/examples/conference-agenda/conference-agenda-example.xml
@@ -1,51 +1,0 @@
-<examples:ExamplePage
-        cssFile="~/examples/conference-agenda/conference-agenda-example.css"
-        xmlns="http://www.nativescript.org/tns.xsd"
-        xmlns:examples="examples/example-base-page"
-        navigatingTo="pageNavigatingTo"
-        tap="tap">
-    <Page.actionBar>
-        <ActionBar title="Conference agenda" automationText="ActionBar">
-            <NavigationButton icon="res://ic_back" tap="goBack" automationText="GoBack" />
-            <Android>
-                <ActionItem id="exampleMenuButton" icon="res://ic_menu_example" automationText="ExampleMenu" />
-            </Android>
-            <iOS>
-                <ActionItem id="exampleMenuButton" ios.position="right" automationText="ExampleMenu">
-                    <ActionItem.actionView>
-                        <Image src="res://ic_menu_example" width="22" height="22" margin="0, -11, 0, 11" />
-                    </ActionItem.actionView>
-                </ActionItem>
-            </iOS>
-        </ActionBar>
-    </Page.actionBar>
-    <GridLayout rows="auto, *" columns="*" class="main-content" id="main-content">
-        <!-- Segmented Bar -->
-        <SegmentedBar row="0" selectedIndex="{{ selectedDay }}"
-                ios:selectedBackgroundColor="white"
-                android:selectedBackgroundColor="#FAC950" 
-                class="conf-segmented-bar">
-            <SegmentedBar.items>
-                <SegmentedBarItem title="MAY 3" />
-                <SegmentedBarItem title="MAY 4" />
-                <SegmentedBarItem title="MAY 5" />
-            </SegmentedBar.items>
-        </SegmentedBar>
-        <GridLayout row="1" rows="auto, *" loaded="onBackgroundLoaded">
-            <!-- Search -->
-            <SearchBar row="0" text="{{ search }}" hint="Search" id="search" tap="inputTap" android:loaded="doNotShowAndroidKeyboard" android:height="40" />
-            <!-- List -->
-            <ListView row="1" items="{{ sessions }}" id="sessions-list" separatorColor="#00000000" ios:itemLoading="changeCellBackground">
-                <ListView.itemTemplate>
-                    <GridLayout columns="auto, auto, *" rows="auto, auto" tap="toggleFavourite" class="session-container" backgroundColor="{{ isFavourite ? '#19FFFFFF' : '#00000000' }}">
-                        <Image row="0" rowSpan="2" col="0" src="{{ isFavourite ? 'res://add_to_fav_1' : 'res://add_to_fav' }}" class="{{ cssClass }}"/>
-                        <Label row="0" col="1" text="{{ range }}" class="session-time"/>
-                        
-                        <Label row="0" col="2" text="{{ room }}" class="session-room"/>                    
-                        <Label row="1" col="1" colSpan="2" text="{{ title }}" textWrap="true" class="session-title"/>
-                    </GridLayout>
-                </ListView.itemTemplate>
-            </ListView>
-        </GridLayout>
-    </GridLayout>
-</examples:ExamplePage>
